pyproject.toml
[tool.poetry]
name = "NoteBook"
version = "0.1.0"
description = ""
authors = ["vitnine <vitninevt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiogram = "^2.10.1"
motor = "^2.3.0"
PyYAML = "^5.3.1"
pydantic = "^1.7.2"
addict = "^2.3.0"
dnspython = "^2.0.0"
Jinja2 = "^2.11.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

